Renamed database_test and created a sqlite script
Hopefully, the script will be able to handle most kinds of input data and store it into a database. The two major functions are adding into database and finding in database. I need to find a way to remove duplicates with the same key. All is based on the user having an associated key that they will use for accessing the database. The database is not meant to store results indefinetly, just for shorter periods of time. However, if I can remove and clear junk from the database, it can be used permanently. NOTE: Might be susceptible for SQL injections. Needs to be fixed