A LOT of changes
Added regions in code for code folding

Updated so that the <tt> tag is changed to the <code> tag when displaying results. Also added <pre> tag to be able to handle multiple spaces in sequence

Added links at the bottom

Added threading, so when the work is queried, the user can do other stuff while its loading. If trying to view before it's done, it will return a page which says loading. I'm currently experimenting with checking the thread and adding a JS to make a loading bar. Not priority

Fixed class based flask.